fix ci bug for deleting data files, test=develop

+++ python/paddle/fluid/contrib/slim/tests/test_post_training_quantization_mobilenetv1.py
@@ -133,8 +133,9 @@ class TestPostTrainingQuantization(unittest.TestCase):
                                                         "small_data", False)
 
         # reader/decorator.py requires the relative path to the data folder
-        cmd = 'rm -rf {0} && ln -s {1} {0}'.format("data",
-                                                   self.data_cache_folder)
+        if not os.path.exists("./data/ILSVRC2012"):
+            cmd = 'rm -rf {0} && ln -s {1} {0}'.format("data",
+                                                       self.data_cache_folder)
         os.system(cmd)
 
         self.batch_size = 1 if os.environ.get('DATASET') == 'full' else 50
